fix: Improve GPU metrics accuracy and process table updates
- Change GPU memory usage from absolute bytes to percentage of total GPU memory
- Add proper command batching for process table updates
- Improve error handling and race condition prevention in collectors
- Fix main.go exit handling to properly kill program on errors

+++ internal/domain/metrics.go
@@ -24,6 +24,6 @@ type GPUMetrics struct {
 type GPUProcessInfo struct {
 	Pid           uint32
 	SmUtil        uint32
-	UsedGpuMemory uint64
+	UsedGpuMemory float64
 	User          string
 }
